requirements.txt
streamlit==1.29.0
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
